Move all DF-related operations to DataFrameOps
