+++ caps-spark/src/main/scala/org/opencypher/caps/impl/spark/SparkSQLExprMapper.scala
@@ -17,13 +17,12 @@ package org.opencypher.caps.impl.spark
 
 import org.apache.spark.sql._
 import org.apache.spark.sql.types._
-import org.opencypher.caps.impl.exception._
 import org.opencypher.caps.api.types._
 import org.opencypher.caps.api.value.CypherValue._
 import org.opencypher.caps.impl.exception.{IllegalArgumentException, IllegalStateException, NotImplementedException}
 import org.opencypher.caps.impl.record.RecordHeader
 import org.opencypher.caps.impl.spark.CAPSFunctions._
-import org.opencypher.caps.impl.spark.convert.SparkUtils._
+import org.opencypher.caps.impl.spark.DataFrameOps._
 import org.opencypher.caps.impl.spark.physical.CAPSRuntimeContext
 import org.opencypher.caps.impl.spark.physical.operators.CAPSPhysicalOperator.columnName
 import org.opencypher.caps.ir.api.expr._
