+++ caps-spark/src/main/scala/org/opencypher/caps/impl/spark/io/hdfs/CsvGraphLoader.scala
@@ -29,7 +29,7 @@ import org.opencypher.caps.api.graph.PropertyGraph
 import org.opencypher.caps.api.io.conversion.{NodeMapping, RelationshipMapping}
 import org.opencypher.caps.api.schema.{CAPSNodeTable, CAPSRelationshipTable}
 import org.opencypher.caps.impl.exception.IllegalArgumentException
-import org.opencypher.caps.impl.spark.convert.SparkUtils.NullabilityOps
+import org.opencypher.caps.impl.spark.DataFrameOps._
 
 trait CsvGraphLoaderFileHandler {
   def location: String
