+++ caps-spark/src/main/scala/org/opencypher/caps/impl/spark/physical/operators/CAPSPhysicalOperator.scala
@@ -24,7 +24,7 @@ import org.opencypher.caps.api.types._
 import org.opencypher.caps.impl.exception.IllegalArgumentException
 import org.opencypher.caps.impl.record.{ColumnName, RecordHeader, RecordSlot, SlotContent}
 import org.opencypher.caps.impl.spark.CAPSConverters._
-import org.opencypher.caps.impl.spark.physical.DataFrameOps._
+import org.opencypher.caps.impl.spark.DataFrameOps._
 import org.opencypher.caps.impl.spark.physical.{CAPSPhysicalResult, CAPSRuntimeContext}
 import org.opencypher.caps.impl.spark.{CAPSGraph, CAPSRecords}
 import org.opencypher.caps.trees.AbstractTreeNode
